Drool website
I improved the interface of the site with React and it became more professional. We can control everything from the admin panel and easily add products to the cart. I also made some features at a higher level, such as API usage in the backend. I also made the filing of the codes more user-friendly.